fix CR comment: edit test instead of duplicating
we want to stop maintaining tests on prev versions
instead make sure all tests run against the latest version
