+++ packages/nouns-contracts/test/governance/NounsDAO/V2/voteRefund.test.ts
@@ -4,8 +4,8 @@ import { solidity } from 'ethereum-waffle';
 import { BigNumber, ContractReceipt } from 'ethers';
 import { ethers } from 'hardhat';
 import {
-  NounsDAOLogicV2,
-  NounsDAOLogicV2__factory,
+  NounsDAOLogicV3__factory,
+  NounsDAOLogicV3,
   NounsDescriptorV2__factory,
   NounsToken,
   Voter__factory,
@@ -14,7 +14,7 @@ import { MaliciousVoter__factory } from '../../../../typechain/factories/contrac
 import {
   address,
   advanceBlocks,
-  deployGovernorV2WithV2Proxy,
+  deployGovernorV3WithV3Proxy,
   deployNounsToken,
   encodeParameters,
   getSigners,
@@ -29,7 +29,7 @@ const { expect } = chai;
 const realLongReason =
   "Judge: The defense may proceed. Roark: Your Honor, I shall call no witnesses. This will be my testimony and my summation. Judge: Take the oath. Court Clerk: Do you swear to tell the truth, the whole truth, and nothing but the truth, so help you God? Roark: I do. Thousands of years ago, the first man discovered how to make fire. He was probably burned at the stake he had taught his brothers to light, but he left them a gift they had not conceived, and he lifted darkness off the earth. Throughout the centuries, there were men who took first steps down new roads, armed with nothing but their own vision. The great creators -- the thinkers, the artists, the scientists, the inventors -- stood alone against the men of their time. Every new thought was opposed; every new invention was denounced. But the men of unborrowed vision went ahead. They fought, they suffered, and they paid. But they won. No creator was prompted by a desire to please his brothers. His brothers hated the gift he offered. His truth was his only motive. His work was his only goal. His work -- not those who used it. His creation -- not the benefits others derived from it -- the creation which gave form to his truth. He held his truth above all things and against all men. He went ahead whether others agreed with him or not, with his integrity as his only banner. He served nothing and no one. He lived for himself. And only by living for himself was he able to achieve the things which are the glory of mankind. Such is the nature of achievement. Man cannot survive except through his mind. He comes on earth unarmed. His brain is his only weapon. But the mind is an attribute of the individual. There is no such thing as a collective brain. The man who thinks must think and act on his own. The reasoning mind cannot work under any form of compulsion. It cannot be subordinated to the needs, opinions, or wishes of others. It is not an object of sacrifice. The creator stands on his own judgment; the parasite follows the opinions of others. The creator thinks; the parasite copies. The creator produces; the parasite loots. The creator's concern is the conquest of nature; the parasite's concern is the conquest of men. The creator requires independence. He neither serves nor rules. He deals with men by free exchange and voluntary choice. The parasite seeks power. He wants to bind all men together in common action and common slavery. He claims that man is only a tool for the use of others -- that he must think as they think, act as they act, and live in selfless, joyless servitude to any need but his own. Look at history: Everything we have, every great achievement has come from the independent work of some independent mind. Every horror and destruction came from attempts to force men into a herd of brainless, soulless robots -- without personal rights, without person ambition, without will, hope, or dignity. It is an ancient conflict. It has another name: \"The individual against the collective.\" Our country, the noblest country in the history of men, was based on the principle of individualism, the principle of man's \"inalienable rights.\" It was a country where a man was free to seek his own happiness, to gain and produce, not to give up and renounce; to prosper, not to starve; to achieve, not to plunder; to hold as his highest possession a sense of his personal value, and as his highest virtue his self-respect. Look at the results. That is what the collectivists are now asking you to destroy, as much of the earth has been destroyed. I am an architect. I know what is to come by the principle on which it is built. We are approaching a world in which I cannot permit myself to live. My ideas are my property. They were taken from me by force, by breach of contract. No appeal was left to me. It was believed that my work belonged to others, to do with as they pleased. They had a claim upon me without my consent -- that it was my duty to serve them without choice or reward. Now you know why a dynamited Courtland. I designed Courtland. I made it possible. I destroyed it. I agreed to design it for the purpose of it seeing built as I wished. That was the price I set for my work. I was not paid. My building was disfigured at the whim of others who took all the benefits of my work and gave me nothing in return. I came here to say that I do not recognize anyone's right to one minute of my life, nor to any part of my energy, nor to any achievement of mine -- no matter who makes the claim! It had to be said: The world is perishing from an orgy of self-sacrificing. I came here to be heard in the name of every man of independence still left in the world. I wanted to state my terms. I do not care to work or live on any others. My terms are: A man's RIGHT to exist for his own sake.";
 const LONG_REASON = realLongReason + realLongReason;
-const REFUND_ERROR_MARGIN = ethers.utils.parseEther('0.00015');
+const REFUND_ERROR_MARGIN = ethers.utils.parseEther('0.001');
 const MAX_PRIORITY_FEE_CAP = ethers.utils.parseUnits('2', 'gwei');
 const DEFAULT_GAS_OPTIONS = { maxPriorityFeePerGas: MAX_PRIORITY_FEE_CAP };
 const MAX_REFUND_GAS_USED = BigNumber.from(200_000);
@@ -39,11 +39,11 @@ let deployer: SignerWithAddress;
 let user: SignerWithAddress;
 let user2: SignerWithAddress;
 let signers: TestSigners;
-let gov: NounsDAOLogicV2;
+let gov: NounsDAOLogicV3;
 let token: NounsToken;
 let snapshotId: number;
 
-describe('Vote Refund', () => {
+describe('V3 Vote Refund', () => {
   before(async () => {
     signers = await getSigners();
     deployer = signers.deployer;
@@ -60,7 +60,7 @@ describe('Vote Refund', () => {
 
     await advanceBlocks(1);
 
-    gov = await deployGovernorV2WithV2Proxy(deployer, token.address);
+    gov = await deployGovernorV3WithV3Proxy(deployer, token.address);
     await submitProposal(user);
   });
 
@@ -353,7 +353,7 @@ describe('Vote Refund', () => {
     // Not using expect emit because it doesn't support the `closeTo` matcher
     // Using longer event parsing because r.events doesn't work when using the Voter contract
     // to simulate multisig usage; events are returned undefined
-    const daoInterface = NounsDAOLogicV2__factory.createInterface();
+    const daoInterface = NounsDAOLogicV3__factory.createInterface();
     const eventId = ethers.utils.id('RefundableVote(address,uint256,bool)');
     const filtered = r.logs.filter(l => l.topics[0] === eventId);
     const parsed = filtered.map(e => {
